+++ test/deserializer/deserialize_inmutable_class_test.dart
@@ -26,7 +26,7 @@ class ImmutableClassInvalidParameter extends _$ImmutableClassInvalidParameterSer
 main() {
   _initMirrors();
 
-  test('deserialize: immutable class 0', () {
+  test('deserialize: immutable class.', () {
     ImmutableClass test = fromJson('{"name":"test", "the_renamed": "test"}', ImmutableClass);
     expect(test.name, equals("test"));
     expect(test.renamed, equals("test"));